Added instruction to fetch repo's xdg-terminal
What we do is, download the ZIP, delete the specific xdg-terminal, and
then download the correct xdg-terminal from the repo. Works great, and
basically ends issue #6, since we're no longer relying on a custom
PyLNP.user any more.

With that, I'm pretty sure now that almost everything in here can be
Licensed to me. Also, I figure, why not tangle the license, just to be
unnecessary and overly dramatic lmao.